performProAPI/api/views.py
from django.shortcuts import render
from api.serializers import EmployeeSerializer
from .models import Employee
from rest_framework.response import Response
from rest_framework.decorators import api_view
# Create your views here.

@api_view(['GET'])
def list_employees(request):
    employees = Employee.objects.all()
    serializer = EmployeeSerializer(employees, many=True)
    return Response(serializer.data)