performProAPI/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.2
importlib_metadata==7.2.0
Markdown==3.6
mysqlclient==2.2.4
sqlparse==0.5.0
typing_extensions==4.12.2
zipp==3.19.2